new code version

+++ Models/TrendsGraphDataInfo.cs
@@ -42,7 +42,7 @@ namespace DataForSeo.Client.Models
         /// <br/>a value of 100 is the peak popularity for the term. A value of 50 means that the term is half as popular. A score of 0 means there was not enough data for this term
         /// </summary>
         [JsonProperty("values", Required = Required.Default, NullValueHandling = NullValueHandling.Ignore)]
-        public IEnumerable<int?> Values { get; set; }
+        public IEnumerable<double?> Values { get; set; }
 
         private IDictionary<string, object> _additionalProperties;
